For the secondary system it will now use the second receive antenna to cancel out interference.

+++ openair_lte/openair1_multi_eNb_UE/PHY/vars.h
@@ -77,7 +77,7 @@ DCI2_5MHz_2A_M10PRB_TDD_t DLSCH_alloc_pdu2;
 
 UE_MODE_t UE_mode;
 
-short **txdataF_rep_tmp;
+int **txdataF_rep_tmp;
 
 #include "PHY/LTE_TRANSPORT/vars.h"
